+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
  - Relax some dependency minimum versions
+ - Add `size_hint()` to some array subset iterators
 
 ## [0.3.0] - 2023-09-27